+++ CHANGELOG.md
@@ -23,3 +23,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   NAudio and AlsaAudio `RecordingDevice` implement. This allows to use to use the same assembly
   on both Windows and Linux (although the limitations what works and what doesn't work remain the
   same)
+
+### Added
+
+- [SIL.Core] new properties for processor architecture: `Platform.ProcessArchitecture` and
+  `Platform.IsRunning64Bit`